+++ CHANGELOG.md
@@ -21,3 +21,4 @@ Created on Dec 20, 2023
 - Beta Release: Transaction deletion and system improvement on December 25, 2023 -> **0.48**
 - Beta Release: Transaction DB model adjustment + deletion UI adjustment on December 26, 2023 -> **0.49**
 - Beta Release: Transaction DB model adjustment + add migration on December 26, 2023 -> **0.491**
+- Beta Release: Changed transaction values on December 26, 2023 -> **0.492**